mitkImageDataItem: changed data type of size member to increase possible image size. Test added
Summary: Merge branch 'T24716-imageDataItem-limit-expanded' fixes T24716

Test Plan: execute mitkImageDataItem-Test

Reviewers: kislinsk, O1 MITK Core

Reviewed By: kislinsk, O1 MITK Core

Maniphest Tasks: T24716

Differential Revision: https://phabricator.mitk.org/D127

+++ Modules/Core/include/mitkImageDataItem.h
@@ -131,7 +131,7 @@ namespace mitk
     bool GetManageMemory() const { return m_ManageMemory; }
     virtual void ConstructVtkImageData(ImageConstPointer) const;
 
-    unsigned long GetSize() const { return m_Size; }
+    size_t GetSize() const { return m_Size; }
     virtual void Modified() const;
 
   protected:
@@ -148,7 +148,7 @@ namespace mitk
 
     bool m_IsComplete;
 
-    unsigned long m_Size;
+    size_t m_Size;
 
   private:
     void ComputeItemSize(const unsigned int *dimensions, unsigned int dimension);
